(fix) issues with windows OS

+++ gegelatilib/src/selector/mapElites/mapElitesSelector.cpp
@@ -38,13 +38,13 @@ std::shared_ptr<Selector::SelectionMetrics> Selector::MapElites::MapElitesSelect
 
 void Selector::MapElites::MapElitesSelector::addArchiveFromDescriptor(size_t nbBins, std::shared_ptr<const MapElitesDescriptor> descriptor, Learn::LearningEnvironment& le)
 {
-    std::pair<size_t, size_t> minAndMaxRange = descriptor->getMinAndMaxRange();
+    std::pair<double, double> minAndMaxRange = descriptor->getMinAndMaxRange();
     mapEliteArchives.insert({descriptor, std::make_shared<MapElitesArchive>(nbBins, descriptor->getNbDescriptors(), minAndMaxRange.first, minAndMaxRange.second)});
 }
 
 void Selector::MapElites::MapElitesSelector::addCvtArchiveFromDescriptor(size_t nbCentroids, std::shared_ptr<const MapElitesDescriptor> descriptor, Learn::LearningEnvironment& le, Mutator::RNG& rng, size_t nbIterationInit, size_t nbDotsInit, double a1, double b1, double a2, double b2)
 {
-    std::pair<size_t, size_t> minAndMaxRange = descriptor->getMinAndMaxRange();
+    std::pair<double, double> minAndMaxRange = descriptor->getMinAndMaxRange();
     mapEliteArchives.insert({descriptor, std::make_shared<CvtMapElitesArchive>(rng, nbCentroids, descriptor->getNbDescriptors(), minAndMaxRange.first, minAndMaxRange.second, nbIterationInit, nbDotsInit, a1, b1, a2, b2)});
 }
